+++ source/Messaging.Api/EventListeners/ActorCreatedListener.cs
@@ -16,6 +16,7 @@ using System;
 using System.Threading.Tasks;
 using Energinet.DataHub.MarketParticipant.Integration.Model.Protobuf;
 using Messaging.Application.Actors;
+using Messaging.Application.Configuration.Commands;
 using Messaging.Infrastructure.Configuration.InternalCommands;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.Logging;
