+++ source/Messaging.Api/EventListeners/ConsumerMovedInListener.cs
@@ -15,6 +15,7 @@
 using System;
 using System.Threading.Tasks;
 using Energinet.DataHub.EnergySupplying.IntegrationEvents;
+using Messaging.Application.Configuration.Commands;
 using Messaging.Application.Transactions.MoveIn;
 using Messaging.Infrastructure.Configuration.InternalCommands;
 using Microsoft.Azure.Functions.Worker;
